ABM Clientes y Planes

+++ back-end/src/main/java/com/obligatorio/obligatorio2dda/service/PlanesViajeService.java
@@ -9,7 +9,7 @@ import com.obligatorio.obligatorio2dda.entity.PlanesViaje;
 @Service
 public interface PlanesViajeService {
     public Iterable<PlanesViaje> findAll();
-    public Optional<PlanesViaje> findById(Long Id);
+    public Optional<PlanesViaje> findById(int Id);
     public PlanesViaje save(PlanesViaje save);
-    public void deleteById(Long Id);
+    public void deleteById(int Id);
 }
